pyproject.toml
[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core"]

[tool.poetry]
authors = ["Georgios Arampatzis <georgios.arampatzis@uoc.gr>"]
description = "Bots for zulip"
name = "zulip-bots"
packages = [
  {include = "bots"}
]
readme = "README.md"
version = "0.1.0"

[tool.poetry.dependencies]
apscheduler = "^3.11.0"
arxiv = "^2.2.0"
docopt = "^0.6.2"
openai = "^1.93.2"
pre-commit = "^4.2.0"
python = "^3.11"
python-dotenv = "^1.1.1"
ratelimit = "^2.2.1"
rich = "^14.0.0"
zulip = "^0.9.0"

[tool.poetry.group.dev.dependencies]
bpython = "^0.25"
pretty-errors = "^1.2.25"

[tool.poetry.scripts]
arxiv = "bots.arxiv.main:main"
kita = "bots.kita.main:main"

[tool.ruff]
exclude = ["sandbox", "docs"]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
ignore = [
  # pylint
  "PLR0913",
  "PLR0915",
  "PLR2004",
  # ruff
  "RUF001",
  # flake8-bandit
  "S101",
  "S102",
  "S602",
  "S603",
  "S604",
  # flake8-annotations
  "ANN",
  # flake8-commas
  "COM819",
  "COM812",
  #
  "NPY002",
  "B027",
  "PD901",
  "PD011",
  "PERF203",
  "A002",
  "A003",
  "PTH",
  "FBT",
  "DTZ",
  "EM",
  "FA100",
  "T201",
  "TRY003",
  "ISC001",
  "N806",
  "N803",
  "SLF001"
]
select = ["ALL"]

[tool.ruff.lint.flake8-quotes]
inline-quotes = "double"
